Move Stream-based interfaces into their own package
